third_party/blink/web_tests/http/tests/devtools/runtime/runtime-es6-setSymbolPropertyValue-expected.txt

Tests editing Symbol properties.


Running: testSetUp
runtime-es6-setSymbolPropertyValue.js:18 ===== Initial =====
runtime-es6-setSymbolPropertyValue.js:19 2
runtime-es6-setSymbolPropertyValue.js:20 

Running: testSetSymbolPropertyValue
runtime-es6-setSymbolPropertyValue.js:18 ===== Set property =====
runtime-es6-setSymbolPropertyValue.js:19 3
runtime-es6-setSymbolPropertyValue.js:20 

Running: testDeleteSymbolProperty
runtime-es6-setSymbolPropertyValue.js:18 ===== Delete property =====
runtime-es6-setSymbolPropertyValue.js:19 undefined
runtime-es6-setSymbolPropertyValue.js:20 
